requirements.txt
requests>=2.31.0
icalendar>=5.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
pymupdf>=1.24.0
openai>=1.30.0
anthropic>=0.25.0
